Refactor contracts
