+++ src/pricecypher/contracts/Script.py
@@ -1,34 +1,14 @@
 from abc import ABC, abstractmethod
-from typing import Any, Optional, Annotated, Callable
+from typing import Any, Optional, Callable
+
+from pricecypher import BaseHandler
 
 
-class Script(ABC):
+class Script(BaseHandler, ABC):
     """
     The abstract Script class serves as an interaction contract such that by extending it with its methods implemented,
         a script can be created that can be used in a generalized yet controlled setting.
     """
-
-    dataset_id: Annotated[int, "The dataset ID"]
-    settings: dict[str, Any]
-    config: dict[str, dict[str, Any]]
-
-    def __init__(self, dataset_id: int, settings: dict[str, Any], config: dict[str, dict[str, Any]]):
-        self.dataset_id = dataset_id
-        self.settings = settings
-        self.config = config
-
-    @abstractmethod
-    def get_config_dependencies(self) -> dict[str, list[str]]:
-        """
-        Fetch the configuration sections and keys in the sections that the script will use that are not yet provided.
-
-        NB: It is not needed to return all required sections and keys, only at least one that has not been provided yet.
-        If all required config is provided, an empty dictionary is to be returned.
-
-        :return: dictionary mapping from section key (string) to a (potentially empty) list of keys of that section
-            that the script requires additionally.
-        """
-        raise NotImplementedError
 
     @abstractmethod
     def get_scope_dependencies(self) -> list[dict[str, Any]]:
@@ -41,6 +21,9 @@ class Script(ABC):
             a 'representation' or a 'scope_id'.
         """
         raise NotImplementedError
+
+    def handle(self, user_input: dict[Any: Any]) -> any:
+        return self.execute('all', self._get_access_token, user_input)
 
     @abstractmethod
     def execute(
